+++ soya/src/client/RegisterClient.js
@@ -1,13 +1,19 @@
 import SoyaClient from './SoyaClient';
 
+// Ensures that webpack hot-reload client does not get loaded more than once.
+if (module.hot && !window.__hotReload) {
+  window.__hotReload = require('webpack-hot-middleware/client');
+}
+
 /**
  * @CLIENT
  * @param {Page} pageClass
  */
 export default function register(pageClass) {
-  console.log('register page called!', pageClass);
+  // TODO: This won't work on older IEs? Logger for client? Log only when debug framework config is set to true?
+  console.log('[REGISTER] Page register', pageClass);
   if (!window.__soyaClient) {
-    console.log('soya client create!');
+    console.log('[REGISTER] Initializing SoyaClient');
     // This follows an implicit contract between renderer and client runtime.
     // Config and RouteArgs must be present as a global variable.
     // Haven't found the better way to do this yet.
@@ -22,7 +28,6 @@ export default function register(pageClass) {
     window.__soyaClient.register(pageClass);
     window.__soyaClient.navigate(pageClass.name, routeArgs, hydratedState);
     return;
-
   }
 
   // TODO: Implement History API page navigation!