+++ soya/src/data/HydrationType.js
@@ -1,7 +1,0 @@
-// TODO: Fix this!
-// Server is always BLOCKING!
-// Client is always NON_BLOCKING!
-// So the correct options would just be LOAD or NOOP.
-export const BLOCKING = 3;
-export const NON_BLOCKING = 2;
-export const NOOP = 1;