+++ soya/src/data/redux/DataComponent.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import ReduxStore from './ReduxStore.js';
+import { isEqualShallow } from './helper.js';
 
 /**
  * Components that wanted to use Soya's redux container may use this.
@@ -8,20 +9,26 @@ import ReduxStore from './ReduxStore.js';
  * @CLIENT_SERVER
  */
 export default class DataComponent extends React.Component {
+  /**
+   * @type {{[key: string]: ActionCreator}}
+   */
+  __soyaActions;
+
+  /**
+   * @type {{[key: string]: Function}}
+   */
+  __soyaUnsubscribe;
+
   constructor(props, context) {
     super(props, context);
-    this.state = {
-      action: {},
-      unsubscribe: {}
-    };
 
     // Both actions and unsubscribe is not part of state because it shouldn't
     // affect rendering in any way. Actions are determined at mounting and
     // shouldn't change throughout this component's lifecycle. Unsubscribe
     // contains functions that unsubscribes this component from state changes,
     // but it's the actual state that matters.
-    this.actions = {};
-    this.unsubscribe = {};
+    this.__soyaActions = {};
+    this.__soyaUnsubscribe = {};
 
     if (!this.getReduxStore() || !this.getReduxStore().__isReduxStore) {
       throw new Error('ReduxStore is not properly wired to this component: ' + this.constructor.name + '.');
@@ -43,13 +50,24 @@ export default class DataComponent extends React.Component {
    * instance. Child components should override this to specify their data
    * requirements.
    *
-   * NOTE: Registration should be done with register() method.
+   * NOTE: Registration must be done with register() method.
+   *
+   * @param {Object} nextProps
    */
-  registerSegments() {
+  registerSegments(nextProps) {
 
   }
 
-  subscribeQuery(props) {
+  /**
+   * Subscribes to queries to a Segment already registered at
+   * registerSegments().
+   *
+   * NOTE: Subscription must be done with subscribe() method. Method must also
+   * use the given props instead of this.props.
+   *
+   * @param {Object} nextProps
+   */
+  subscribeQueries(nextProps) {
 
   }
 
@@ -58,7 +76,7 @@ export default class DataComponent extends React.Component {
    */
   register(segment) {
     var actionCreator = this.getReduxStore().register(segment);
-    this.actions[segment._getName()] = actionCreator;
+    this.__soyaActions[segment._getName()] = actionCreator;
   }
 
   /**
@@ -69,8 +87,10 @@ export default class DataComponent extends React.Component {
    * @param {?Object} hydrationOption
    */
   subscribe(segmentName, query, stateName, queryOptions, hydrationOption) {
-    // Unregister if already registered.
-    this.unregister(stateName);
+    // Unsubscribe if already subscribed.
+    this.unsubscribe(stateName);
+
+    console.log('[DATA] Subscribe', this, segmentName);
     var callback = (newState) => {
       this.setState({[stateName]: newState});
     };
@@ -86,29 +106,62 @@ export default class DataComponent extends React.Component {
   /**
    * @param {string} stateName
    */
-  unregister(stateName) {
-    delete this.state[stateName];
-    if (this.unsubscribe[stateName]) {
-      this.unsubscribe[stateName]();
-      delete this.unsubscribe[stateName];
+  unsubscribe(stateName) {
+    console.log('[DATA] Un-register', this, stateName);
+    if (this.__soyaUnsubscribe[stateName]) {
+      this.__soyaUnsubscribe[stateName]();
+      delete this.__soyaUnsubscribe[stateName];
     }
   }
 
-  componentWillReceiveProps(nextProps) {
+  /**
+   * This implementation assumes:
+   *
+   * 1) All DataComponent(s) render only using props and states (i.e. they
+   *    are pure render components). Child classes can override this method
+   *    if they are not.
+   * 2) All DataComponent(s) parents send their props as immutable objects
+   *    (i.e. if the prop changes, they recreate the whole object).
+   * 3) All DataComponent(s) set their state as immutable objects
+   *    (i.e. if state changes, they recreate the whole object).
+   *
+   * Based on the above assumptions, this method will do a shallow comparison of
+   * both state and props. If the above assumptions are not correct, user can
+   * override this method.
+   */
+  shouldComponentUpdate(nextProps, nextState) {
+    var shouldUpdate = !isEqualShallow(this.props, nextProps) || !isEqualShallow(this.state, nextState);
+    console.log('[DATA] Should update?', this, shouldUpdate);
+    return shouldUpdate;
+  }
 
+  componentWillUpdate(nextProps, nextState) {
+    console.log('[DATA] Will update', this, nextProps, nextState);
+  }
+
+  /**
+   * @param {Object} nextProps
+   */
+  componentWillReceiveProps(nextProps) {
+    // TODO: Logger at client! Remove if debug is set to false!
+    console.log('[DATA] Receive new props', this, nextProps);
+    this.subscribeQueries(nextProps);
   }
 
   /**
    * Registers the store. Run at both client and server side when rendering.
    */
   componentWillMount() {
-    this.registerSegments();
+    console.log('[DATA] Mounting', this, this.props);
+    this.registerSegments(this.props);
+    this.subscribeQueries(this.props);
   }
 
   /**
-   * Unregister all segments.
+   * Unsubscribe all segments.
    */
   componentWillUnmount() {
+    console.log('[DATA] Unmounting', this);
     this.getReduxStore().unsubscribe(this);
   }
 }